Move revenue table to its own component
